services: bump record read_all max_records to 150
We have 100+ resource types at NU so that limit constantly crashes selection
of certain resource types. We have other vocabularies that have the potential
to grow past 100 as well.

Tracking every use of read_all is unfeasible so making the change here.

Until we figure out a better way to address these instance-specific
scaling concerns, I am bumping the default max_records so that
this works for us and probably others.