service: use "read" permission in `read_many`
In most of the `RecordsService.create_search` calls throughout the
codebase, we're passing `permission_action="read"`. This is because
`can_read` is the permission where we describe in each generator's
`Generator.query_filter` the search filtering logic which targets
the individual entities that an `identity` can access, for example
"my records", "records in communities I'm a curator", etc.

The only case where we don't use `read`, is when we use `read_deleted`,
which is usually just stricter version of `read`.

Relying on the `(can_)search` permission is usually wrong, since this is
used to check if someone "can" search, but not "what" they can search.