feat(records): return page number in RecordList
* Added the current page number as part of the serialized RecordList
result type. This is needed for cases where the page number included in
the request differs from the actual page the view handler decided to
return.
* Specifically right now we need this for
https://github.com/inveniosoftware/invenio-requests/issues/494, where
the JS client requests a page that includes a given record (rather than
a specific numbered page), and then needs to know which page was
actually returned so it can update the local state.